Fix empty file test in database init
"wwinit database" was failing when creating a new database because the
test using "wc" was failing with syntax error in Bash. Since what we
want to know is that the file exists and has something in it, a simple
"-s" should suffice.

Tested locally and it passed that point, whereas before it was failing
with error:

/usr/libexec/warewulf/wwinit/10-database.init: line 167:
[: too many arguments
Lethal error thrown by module:
/usr/libexec/warewulf/wwinit/10-database.init