Generate RSA ssh keys for cluster access instead of DSA
OpenSSH 7.0 disabled support for DSA keys by default. While support
can be reenabled via configuration, they were removed due to security
concerns, so it seems safer to switch to another algorithm. Ed25519
might in some sense be the best, but it's quite new and might thus not
yet be available on all distros Warewulf supports. So use RSA which is
good enough for now.

http://www.openssh.com/txt/release-7.0
http://www.openssh.com/legacy.html
https://meyering.net/nuke-your-DSA-keys/
https://security.stackexchange.com/questions/5096/rsa-vs-dsa-for-ssh-authentication-keys