vnfs: Add --auto-agree-with-licenses to include-suse PKGR_CMD
When attempting to create an opensuse chroot, the wwmkchroot script
would fail on packages that attempt to display a license and expect a
response from the console.

The zypper package manager has an option that automatically accepts
package license agreements when present. Adding this option to the
PKGR_CMD variable resolved the issue with mkchroot.

Signed-of-by: John L. Jolly <jjolly@suse.com>