Suse prov config local binary copy (#140)
* provision: Prepare configure.ac for local file options

The FIND_LOCAL macros is created in the configure.ac to handle the
selection of locally installed binaries as alternatives to the
Warewulf-built binaries. By default, the macro does not set any
variables and makes no changes if the --with-local-* option is not used
or is used as --without-local-*.

The FIND_LOCAL macro takes two parameters:
* package name: This parameter is used for three purposes:
  - It is used for the configure script option in the form of
    --with-local-<package name>
  - It sets a display string shown during the run of the configure
    script
  - It is used to form the AC_SUBST variable local_<package name>_path
     that can be used in *.in files
* file name: This parameter is the file to search for. If the file is
  found, then the local_<package name>_path variable is set to the full
  path of the file. If the file cannot be found, a command with the file
  name will be searched for using the `which` command.

Signed-off-by: John L. Jolly <jjolly@suse.com>

* provision: allow build to use locally-installed busybox

Traditionally, distribution vendors have their own build of the Busybox
utility. Having Warewulf build a specialized copy of Busybox causes
issues:
* Duplicate builds cause unnecessary extra time and resources to be
  spent in the build process
* Patches need to be applied and tested twice

To mitigate this issue, two extra configuration items have been added
to the provision configure script:
* --with-local-busybox allows the builder to specify the location of
  the busybox binary
  - If no path to the binary is specified, then the PATH is searched
    for a suitable binary
  - If a suitable binary cannot be found, the system falls back to
    building Busybox internally
* --with-busybox-links-file specifies the list of links that need to be
  created along with the Busybox binary
  - If a local Busybox binary is found, this defaults to
    /usr/share/busybox/busybox.links
  - This file must exist for the local Busybox binary to be used

In all cases, if these configuration items are not used, the build
falls back to building Busybox internally.

Signed-off-by: John Jolly <jjolly@suse.com>

* provision: Add build configuration to allow for use of local e2fsprogs

In order to shorten the build process as well as allow distributions
the ability to use a pre-built binary, a configure options has been
added that allows the use of a local installation of the e2fsprogs
utility mkfs.ext4.

Signed-off-by: John L. Jolly <jjolly@suse.com>

* provision: Add build configuration to allow for use of local ipxe

In order to shorten the build process as well as allow distributions
the ability to use a pre-built binary, a configure options has been
added that allows the use of a local installation of the ipxe images
undionly.kpxe, snp-i386.efi, and snp-x86_64.efi.

Signed-off-by: John L. Jolly <jjolly@suse.com>

* provision: Add build configuration to allow for use of local bsdtar

In order to shorten the build process as well as allow distributions
the ability to use a pre-built binary, a configure options has been
added that allows the use of a local installation of the libarchive
file bsdtar.

Signed-off-by: John L. Jolly <jjolly@suse.com>

* provision: Add build configuration to allow for use of local parted

In order to shorten the build process as well as allow distributions
the ability to use a pre-built binary, a configure options has been
added that allows the use of a local installation of the parted files
parted and partprobe.

Signed-off-by: John L. Jolly <jjolly@suse.com>