Remove shebang from scripts only intended to be sourced
When a shell script is not intended to be executed as a stand-alone
executable, the shebang at the start of the script is not necessary. In
cases where the script should not be executed, the shebang should be
removed. The following scripts are specifically sourced by other
executable scripts and thus have no need for the shebang:

    cluster/etc/cluster-env.csh
    cluster/etc/cluster-env.sh
    vnfs/libexec/wwmkchroot/functions

Signed-off-by: John L. Jolly <jjolly@suse.com>