Add in missing logic for setting VNFS name in `wwsh vnfs set` (#168)
For whatever reason, in the second where command eq "set", only $opt_chroot or $opt_arch could be set with wwsh vnfs set -- the logic was missing. I added it in, taking the logic from the $opt_arch section. Working as expected, and does not negatively impact wwsh import functionality.