provision: Fix setting of $prefix when no --prefix is specified
   Commit 98bd7a371482e2082ce08f34a95725387aa90be9
   Author: jcsiadal <jeremy.c.siadal@intel.com>
   Date:   Wed Feb 12 13:11:09 2020 -0800

   WW Provision working

attempted to fix the rather myserious way $prefix was handled
in this subproject. Unfortunately, the fix did not quite succeed:
The macro used: 'AC_PREFIX_DEFAULT()' doesn't place code at the
location where it is used but but into the 'default' section of
generated script using: 'm4_divert_text([DEFAULTS], ..]'
At this location $WAREWULF_PREFIX is not yet defined, thus the
variable name - not it's value needs to be assigned.
'AC_EXPAND_DIR()' is then able to expand this name - provided
it is set before this macro is expanded.

Signed-off-by: Egbert Eich <eich@suse.com>