Add /lib/modules/$opt_kversion/sysctl.conf to initfs if present
Kernels may ship their own sysctl.conf with settings specific
to the respective kernel version. This is shipped with the
kernel itself and - due to the 'UsrMerge'-efforts may be installed
in the module directory instead of /boot/sysctl.conf-${KVERSION}.

Link it to there if it doesn't exist there already.

Signed-off-by: Egbert Eich <eich@suse.com>