Add Support for Zstandard Compressed Kernel Modules
Newer distros (eg. OpenSUSE 15.4) are shipping with zstd compressed
kernel modules. Currently busybox does not have support for zstd.
Thus we add a new configure argument in provision `--with-local-kmod`
to pull in kmod from the build host to the initramfs and symlink it to
modprobe, etc. When this option is used we drop the module
utilites from the busybox config to save a bit of space.

Also update wwbootstrap to look for zst suffix kernel modules, and
update the check_compression logic a bit to look for each compression
type.